Initial commit of semidbm, a semi-dumb dbm
The intent of semidbm is to represent a slight improvement
to the existing dumbdbm of the standard library.  It is
a pure python library, intended to be cross platform, and
to only be semi-dumb dbm, instead of completely dumb dbm.

This initial version is missing a lot of functionality,
and only supports __setitem__ and __getitem__.  I included
a basic profiling script that compares the various dbms.  The
initial tests show that semi-dumb dbm performs slightly better
than dumbdbm.