Add support for key deletion
This uses a pretty simple scheme, a value
of -1 for the value length is used to indicate
that a key has been deleted.  When the index is
read in, if it encounters a key with value length
-1, it will delete they key from the in memory
index.

There is one problem that hasn't been addressed,
and it's that both the index and the data file
continually grow.  The index can be compacted
when it's first read in or when it's closed.
Same with the data file.  But for now, key deletions!