Add read only mmap implementation
This was faster in all the read benchmarks I ran.
This is currently the default implementation if opened
open(..., 'r').  I've left the original, non-mmapped version
in the code in case the mmap version has issues for
people.  You have to directly instantiate it via
_SemiDBMReadOnly(...) for the time being.  I'm hoping that
eventually I can just remove _SemiDBMReadOnly altogether, but
the mmap version needs more testing first.