Rewrite benchmark script
I have no idea what a reasonable benchmark
looks like so I figured it was best to see
what smarter people have used as a benchmark
and copy them.

This rewrite of the benchmark script is
heavily based on google's leveldb benchmark
(db_bench).  It performs similar benchmarks
and outputs similar stats. Also the same
default values used for the leveldb
benchmark are used for this benchmark,
which is basically 1 million entries with
16 byte keys and 100 byte values.  These
parameters are configurable so it's
easy to benchmark with larger values.

And that being said, semidbm actually
beats gdbm in most of the benchmarks.
I'm going to try to get other dbms installed
so I can have more data points.