Use non mmaped version for read only mode
I'm reluctant to do this, but I using the non
mmapped version gives very predictable performance.
The mmaped version is in every way faster than
lseek version, but the performance degrades as
the key size increases.

I've kept the code in the semidbm module, but
semidbm.open(..., 'r') will use the non mmapped
version for now.  I might eventually revert this
in the future, but the mmapped version needs more
research.