Move the db files to a separate db
Previously you'd have to specify the name of the
data file to open the db (it would assume the index file
was just datafile + ".idx".

Now the "filename" is actually a dirname:

mydb
|
|
|_____ data
|
|_____ data.idx

To load this you'd say:

semidbm.open('mydb', 'r')

This gives me more flexibility in the future and it keeps
all the data files in a specific db.

This is backwards incompatible with <= 0.2.1, so I've already
bumped the version to 0.3.0.  It's pretty easy to migrate though,
in the example above, just mkdir "mydb", and move the data file
to "mydb/data", and the index file to "mydb/data.idx"