Add windows support
This required a minor refactoring of the code.  os.open is used
instead of __builtin__.open for writing to the index/data file (though
for loading/compacting __builtin__.open is still used since it works
as expected on windows).

Also, windows does not support atomic renames so a _Renamer abstraction
was added to support windows while still letting posix clients have
atomic renames.

Windows also does not have a compatible mmap interface, so the
access kwarg is used which provides consistent behavior for windows
and posix platforms.

And one final thing that was needed was that you cannot modify
(delete/rename/etc) a file is anything else currently has it opened.
However, on posix, this is perfectly valid.  In this case, I just
made sure to close all the files before renaming/removing them.

Unittests pass on win7 running python 2.7.2.