Use bytes instead of str for benchmarks
So bsddb3 actually just requires bytes for key/values
so this is needed if we want to benchmark bsddb3.  Though
if I could remove the isinstance checks in semidbm, I'd get
a small performance boost as well ;)

The benchmarks for bsddb3 and semidbm are fairly close for
1000000 keys, though for smaller key sizes bsddb3 is much
faster than semidbm.